# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Confirmation": {
        "additionalProperties": false,
        "properties": {
          "final_answer": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Answer"
          },
          "transcription": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Transcription",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "transcription"
        ],
        "title": "Confirmation",
        "type": "object"
      },
...
      "FeedbackPublic": {
        "additionalProperties": false,
        "description": "Historical feedback records absent structured observations as unknown.",
        "properties": {
          "concepts": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Concepts",
            "type": "array"
          },
          "guidance": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Guidance",
            "type": "array"
          },
          "learning_observation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LearningObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_step": {
            "maxLength": 1000,
            "title": "Next Step",
            "type": "string"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Strengths",
            "type": "array"
          },
          "teaching_action": {
            "anyOf": [
              {
                "enum": [
                  "acknowledge",
                  "clarify",
                  "explain",
                  "coach",
                  "extend"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teaching Action"
          },
          "uncertainty_note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertainty Note"
          }
        },
        "required": [
          "strengths",
          "guidance",
          "next_step",
          "concepts",
          "teaching_action",
          "learning_observation"
        ],
        "title": "FeedbackPublic",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearningObservation": {
        "additionalProperties": false,
        "description": "Fallible evidence about this response, never a mastery score or a grade.",
        "properties": {
          "assessment": {
            "enum": [
              "sufficient",
              "developing",
              "uncertain",
              "not_assessed"
            ],
            "title": "Assessment",
            "type": "string"
          },
          "evidence": {
            "maxLength": 600,
            "minLength": 1,
            "title": "Evidence",
            "type": "string"
          },
          "open_points": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 3,
            "title": "Open Points",
            "type": "array"
          },
          "resolved_points": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 3,
            "title": "Resolved Points",
            "type": "array"
          }
        },
        "required": [
          "assessment",
          "evidence",
          "resolved_points",
          "open_points"
        ],
        "title": "LearningObservation",
        "type": "object"
      },
...
      "OperationPublic": {
        "properties": {
          "ambiguities": {
            "items": {
              "type": "string"
            },
            "title": "Ambiguities",
            "type": "array"
          },
          "assistance_level": {
            "default": 0,
            "title": "Assistance Level",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interpretation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpretation"
          },
          "interpretation_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpretation Version"
          },
          "interpreted_final_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpreted Final Answer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "problem_id": {
            "format": "uuid",
            "title": "Problem Id",
            "type": "string"
          },
          "reading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadingPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "safe_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "work_text": {
            "title": "Work Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "problem_id",
          "kind",
          "text",
          "status",
          "work_text",
          "safe_error",
          "created_at"
        ],
        "title": "OperationPublic",
        "type": "object"
      },
...
      "ReadingPublic": {
        "additionalProperties": false,
        "properties": {
          "ambiguities": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Ambiguities",
            "type": "array"
          },
          "can_continue": {
            "title": "Can Continue",
            "type": "boolean"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "organization_feedback": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Organization Feedback",
            "type": "array"
          },
          "quality": {
            "enum": [
              "clear",
              "uncertain",
              "unreadable"
            ],
            "title": "Quality",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "transcription": {
            "maxLength": 8000,
            "title": "Transcription",
            "type": "string"
          }
        },
        "required": [
          "transcription",
          "quality",
          "confidence",
          "ambiguities",
          "organization_feedback",
          "rejection_reason",
          "can_continue"
        ],
        "title": "ReadingPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerdictPublic": {
        "properties": {
          "answer_status": {
            "title": "Answer Status",
            "type": "string"
          },
          "format_status": {
            "title": "Format Status",
            "type": "string"
          },
          "reasoning_status": {
            "title": "Reasoning Status",
            "type": "string"
          },
          "verifier_version": {
            "title": "Verifier Version",
            "type": "string"
          }
        },
        "required": [
          "answer_status",
          "format_status",
          "reasoning_status",
          "verifier_version"
        ],
        "title": "VerdictPublic",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shepherd Academy Universe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/problems/{problem_id}/photos": {
      "post": {
        "operationId": "upload_photo_api_v1_problems__problem_id__photos_post",
        "parameters": [
          {
            "in": "path",
            "name": "problem_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Problem Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "answer",
              "enum": [
                "answer",
                "question"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Photo",
        "tags": [
          "photos"
        ]
      }
    },
...
    "/api/v1/submissions/{submission_id}/confirm-interpretation": {
      "post": {
        "operationId": "confirm_api_v1_submissions__submission_id__confirm_interpretation_post",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Confirmation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/submissions/{submission_id}/image": {
      "get": {
        "operationId": "photo_api_v1_submissions__submission_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Photo",
        "tags": [
          "photos"
        ]
      }
    },
...
  }
}